pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = [
    {name = "AlexShcherb1173",email = "Alex_181173@list.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "deepdiff (>=8.5.0,<9.0.0)",
    "pandas-stubs (>=2.3.0.250703,<3.0.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "google-api-python-client (>=2.181.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "docker (>=7.1.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
mypy = "^1.16.0"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
flake8 = "^7.3.0"
mypy = "^1.17.1"
black = "^25.1.0"
isort = "^6.0.1"
requests = "^2.32.5"

[tool.black]
line-length = 119
exclude = '''/\.git/'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '''/\.venv/'''
